Make KzgDecidingKey fields public
Fixes #55.

This commit enables users to re-use more of snark-verifier code when
implementing

  impl AccumulationDecider<_, MyLoader> for KzgAs<_, _>

in order to define their own Loader and make it work with KZG10.

The field `svk` was already exported through `impl
AsRef<KzgSuccinctVerifyingKey<M::G1Affine>>`, but this commit
additionally defines it as `pub` for consistency with `g2` and `s_g2`
fields.